\name{setup_skywatch_oauth}
\alias{setup_skywatch_oauth}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Sets up the OAuth credentials for an SkyWatch session
}
\description{
Sets up the OAuth credentials for an SkyWatch session
}
\usage{
setup_skywatch_oauth(api_key)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{api_key}{
api_key
}
}
\details{
details
}
\value{
output
}


%% ~Make other sections like Warning with \section{Warning }{....} ~

\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x) 
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ spatial }% use one of  RShowDoc("KEYWORDS")

