\name{SkyWatchr-package}
\alias{SkyWatchr}
\alias{SkyWatchr-package}
\docType{package}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Wrapper for the SkyWatch API
}
\description{
Wrapper for the SkyWatch API.
    It queries and download satellite imagery using the SkyWatch API.
}
\details{
\tabular{ll}{
Package: \tab SkyWatchr\cr
Type: \tab Package\cr
Version: \tab 0.3-0\cr
Date: \tab 2016-09-24\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ali Santacruz, SkyWatch API Developers

Maintainer: Ali Santacruz <asantacruzdelgado@clarku.edu>
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ package }% use one of  RShowDoc("KEYWORDS")
\keyword{ spatial }% __ONLY ONE__ keyword per line
