\name{SkyWatchr-package}
\alias{SkyWatchr}
\alias{SkyWatchr-package}
\docType{package}
\title{
Wrapper for the SkyWatch API
}
\description{
Query and download satellite imagery and climate/atmospheric datasets using the SkyWatch API. 
     Search datasets by wavelength (band), cloud cover, resolution, location, date, etc.
     Get the query results as data frame and as HTML. 
     
     To learn more about the SkyWatch API, see \href{https://github.com/skywatchspaceapps/api}{https://github.com/skywatchspaceapps/api}. An API key for free access to the SkyWatch API can be requested at \href{http://www.skywatch.co/request-access}{http://www.skywatch.co/request-access}.
}
\details{
\tabular{ll}{
Package: \tab SkyWatchr\cr
Type: \tab Package\cr
Version: \tab 0.5-1\cr
Date: \tab 2017-01-08\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Ali Santacruz, SkyWatch API Developers

Maintainer: Ali Santacruz <amsantac@unal.edu.co>
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ package }% use one of  RShowDoc("KEYWORDS")
\keyword{ spatial }% __ONLY ONE__ keyword per line
